Add support for inputs of type string
We currently only support BigDecimal and other Numerics, however, it
is also desirable to support numerics represented as Strings. We can
easily do that by calling to_f on the input to always turn it into a
Float.

In order to verify turning inputs into Floats is always correct, we
added a test to compare the results of formatting via Ruby and our
implementation.

Knowing that we always produce the right result for Floats, we can
simplify our implementation to skip the check for BigDecimal and always
convert the value to a Float. This further improves the performance
by about 10%.